try
funcional

+++ Proyecto_p04/Proyecto_p04/frmVentanilla2.cs
@@ -89,30 +89,33 @@ namespace Proyecto_p04
             {
                 DialogResult result = MessageBox.Show("¿Desea guardar los datos?", "Confirmar Guardar", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
-                //Codigo Guardar
-                conexionBD.conectarBD();
-                string insertar = "INSERT INTO tbl_ventanilla(Id,Nombre,Identificacion,HoraFecha,Usuario,Destino,Asiento,NumeroV,Boleto,Nacionalidad,NPasaporte,HoraEntrada,HoraSalida) VALUES(@Id,@Nombre,@Identificacion,@HoraFecha,@Usuario,@Destino,@Asiento,@NumeroV,@Boleto,@Nacionalidad,@NPasaporte,@HoraEntrada,@HoraSalida)";
-                SqlCommand cmd = new SqlCommand(insertar, conexionBD.conectarBD());
-                cmd.Parameters.AddWithValue("@Id", txtId.Text);
-                cmd.Parameters.AddWithValue("@Nombre", txtNombre.Text);
-                cmd.Parameters.AddWithValue("@Identificacion", txtIdentificacion.Text);
-                cmd.Parameters.AddWithValue("@HoraFecha", txtHoraFecha.Text);
-                cmd.Parameters.AddWithValue("@Usuario", txtUsuario.Text);
-                cmd.Parameters.AddWithValue("@Destino", txtVuelo.Text);
-                cmd.Parameters.AddWithValue("@NumeroV", txtNumeroVentanilla.Text);
-                cmd.Parameters.AddWithValue("@Boleto", txtBoleto.Text);
-                cmd.Parameters.AddWithValue("@Nacionalidad", cbNacionalidad.SelectedIndex);
-                cmd.Parameters.AddWithValue("@NPasaporte", txtNPasaporte.Text);
-                cmd.Parameters.AddWithValue("@HoraEntrada", txtEntrada.Text);
-                cmd.Parameters.AddWithValue("@HoraSalida", txtSalida.Text);
+                if (result == DialogResult.Yes)
+                {
+                    //Codigo Guardar
+                    conexionBD.conectarBD();
+                    string insertar = "INSERT INTO tbl_ventanilla(Id,Nombre,Identificacion,HoraFecha,Usuario,Destino,Asiento,NumeroV,Boleto,Nacionalidad,NPasaporte,HoraEntrada,HoraSalida) VALUES(@Id,@Nombre,@Identificacion,@HoraFecha,@Usuario,@Destino,@Asiento,@NumeroV,@Boleto,@Nacionalidad,@NPasaporte,@HoraEntrada,@HoraSalida)";
+                    SqlCommand cmd = new SqlCommand(insertar, conexionBD.conectarBD());
+                    cmd.Parameters.AddWithValue("@Id", txtId.Text);
+                    cmd.Parameters.AddWithValue("@Nombre", txtNombre.Text);
+                    cmd.Parameters.AddWithValue("@Identificacion", txtIdentificacion.Text);
+                    cmd.Parameters.AddWithValue("@HoraFecha", txtHoraFecha.Text);
+                    cmd.Parameters.AddWithValue("@Usuario", txtUsuario.Text);
+                    cmd.Parameters.AddWithValue("@Destino", txtVuelo.Text);
+                    cmd.Parameters.AddWithValue("@NumeroV", txtNumeroVentanilla.Text);
+                    cmd.Parameters.AddWithValue("@Boleto", txtBoleto.Text);
+                    cmd.Parameters.AddWithValue("@Nacionalidad", cbNacionalidad.SelectedIndex);
+                    cmd.Parameters.AddWithValue("@NPasaporte", txtNPasaporte.Text);
+                    cmd.Parameters.AddWithValue("@HoraEntrada", txtEntrada.Text);
+                    cmd.Parameters.AddWithValue("@HoraSalida", txtSalida.Text);
 
-                cmd.Parameters.AddWithValue("@Asiento", control.ToString()); //esta es la variable para los nombres de los chbt
+                    cmd.Parameters.AddWithValue("@Asiento", control.ToString()); //esta es la variable para los nombres de los chbt
 
-                cmd.ExecuteNonQuery();
+                    cmd.ExecuteNonQuery();
 
 
-                MessageBox.Show("Los datos fueron agregados de forma exitosa!!!");
-                dataGridView1.DataSource = LLenar_grid1();
+                    MessageBox.Show("Los datos fueron agregados de forma exitosa!!!");
+                    dataGridView1.DataSource = LLenar_grid1();
+                }
             }
             catch (Exception ex)
             {
@@ -124,33 +127,36 @@ namespace Proyecto_p04
         {
             try
             {
-                DialogResult result = MessageBox.Show("¿Desea modificar los datos?", "Confirmar Modificacion", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                DialogResult result = MessageBox.Show("¿Desea guardar los datos?", "Confirmar Guardar", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
-                //Codigo Modificar
+                if (result == DialogResult.Yes)
+                {
+                    //Codigo Modificar
 
-                conexionBD.conectarBD();
-                string actualizar = "UPDATE tbl_ventanilla SET Id=@Id,Nombre=@Nombre,Identificacion=@Identificacion,HoraFecha=@HoraFecha,Usuario=@Usuario,Destino=@Destino,NumeroV=@NumeroV,Asiento=@Asiento,Boleto=@Boleto,Nacionalidad=@Nacionalidad,NPasaporte=@NPasaporte,HoraEntrada=@HoraEntrada,HoraSalida=@HoraSalida" +
-                "WHERE Id=@Id";
-                SqlCommand cmd = new SqlCommand(actualizar, conexionBD.conectarBD());
-                cmd.Parameters.AddWithValue("@Id", txtId.Text);
-                cmd.Parameters.AddWithValue("@Nombre", txtNombre.Text);
-                cmd.Parameters.AddWithValue("@Identificacion", txtIdentificacion.Text);
-                cmd.Parameters.AddWithValue("@HoraFecha", txtHoraFecha.Text);
-                cmd.Parameters.AddWithValue("@Usuario", txtUsuario.Text);
-                cmd.Parameters.AddWithValue("@Destino", txtVuelo.Text);
-                cmd.Parameters.AddWithValue("@NumeroV", txtNumeroVentanilla.Text);
-                cmd.Parameters.AddWithValue("@Boleto", txtBoleto.Text);
-                cmd.Parameters.AddWithValue("@Nacionalidad", cbNacionalidad.SelectedIndex);
-                cmd.Parameters.AddWithValue("@NPasaporte", txtNPasaporte.Text);
-                cmd.Parameters.AddWithValue("@HoraEntrada", txtEntrada.Text);
-                cmd.Parameters.AddWithValue("@HoraSalida", txtSalida.Text);
+                    conexionBD.conectarBD();
+                    string actualizar = "UPDATE tbl_ventanilla SET Id=@Id,Nombre=@Nombre,Identificacion=@Identificacion,HoraFecha=@HoraFecha,Usuario=@Usuario,Destino=@Destino,NumeroV=@NumeroV,Asiento=@Asiento,Boleto=@Boleto,Nacionalidad=@Nacionalidad,NPasaporte=@NPasaporte,HoraEntrada=@HoraEntrada,HoraSalida=@HoraSalida" +
+                    "WHERE Id=@Id";
+                    SqlCommand cmd = new SqlCommand(actualizar, conexionBD.conectarBD());
+                    cmd.Parameters.AddWithValue("@Id", txtId.Text);
+                    cmd.Parameters.AddWithValue("@Nombre", txtNombre.Text);
+                    cmd.Parameters.AddWithValue("@Identificacion", txtIdentificacion.Text);
+                    cmd.Parameters.AddWithValue("@HoraFecha", txtHoraFecha.Text);
+                    cmd.Parameters.AddWithValue("@Usuario", txtUsuario.Text);
+                    cmd.Parameters.AddWithValue("@Destino", txtVuelo.Text);
+                    cmd.Parameters.AddWithValue("@NumeroV", txtNumeroVentanilla.Text);
+                    cmd.Parameters.AddWithValue("@Boleto", txtBoleto.Text);
+                    cmd.Parameters.AddWithValue("@Nacionalidad", cbNacionalidad.SelectedIndex);
+                    cmd.Parameters.AddWithValue("@NPasaporte", txtNPasaporte.Text);
+                    cmd.Parameters.AddWithValue("@HoraEntrada", txtEntrada.Text);
+                    cmd.Parameters.AddWithValue("@HoraSalida", txtSalida.Text);
 
-                //cmd.Parameters.AddWithValue("@Asiento",  control); //esta es la variable para los nombres de los chbt
+                    //cmd.Parameters.AddWithValue("@Asiento",  control); //esta es la variable para los nombres de los chbt
 
 
-                cmd.ExecuteNonQuery();
-                MessageBox.Show("Los datosfueron agregados de forma exitosa!!!");
-                dataGridView1.DataSource = LLenar_grid1();
+                    cmd.ExecuteNonQuery();
+                    MessageBox.Show("Los datosfueron agregados de forma exitosa!!!");
+                    dataGridView1.DataSource = LLenar_grid1();
+                }
             }
             catch (Exception ex)
             {
@@ -162,17 +168,20 @@ namespace Proyecto_p04
         {
             try
             {
-                DialogResult result = MessageBox.Show("¿Desea eliminar los datos?", "Confirmar solicitud de eliminar", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+                DialogResult result = MessageBox.Show("¿Desea guardar los datos?", "Confirmar Guardar", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
-                //Codigo Eliminar
-                conexionBD.conectarBD();
-                string eliminar = "DELETE FROM tbl_ventanilla WHERE Id=@Id";
-                SqlCommand cmd = new SqlCommand(eliminar, conexionBD.conectarBD());
-                cmd.Parameters.AddWithValue("@Id", txtId.Text);
-                cmd.ExecuteNonQuery();
+                if (result == DialogResult.Yes)
+                {
+                    //Codigo Eliminar
+                    conexionBD.conectarBD();
+                    string eliminar = "DELETE FROM tbl_ventanilla WHERE Id=@Id";
+                    SqlCommand cmd = new SqlCommand(eliminar, conexionBD.conectarBD());
+                    cmd.Parameters.AddWithValue("@Id", txtId.Text);
+                    cmd.ExecuteNonQuery();
 
-                MessageBox.Show("Los datos fueron Eliminados!!!");
-                dataGridView1.DataSource = LLenar_grid1();
+                    MessageBox.Show("Los datos fueron Eliminados!!!");
+                    dataGridView1.DataSource = LLenar_grid1();
+                }
             }
             catch (Exception ex)
             {
@@ -183,16 +192,8 @@ namespace Proyecto_p04
         private void btnLimpiar_Click(object sender, EventArgs e)
         {
 
-            try
-            {
-                DialogResult result = MessageBox.Show("¿Desea limpiar los campos?", "Confirmar accion de limpieza", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-
                 limpiar_txt();
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show($"Ocurrió un error: {ex.Message}", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
-            }
+            
         }
 
         public void paises()
